% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_search.R
\name{attribute_finder}
\alias{attribute_finder}
\title{Attribute Finder}
\usage{
attribute_finder(xml_list_obj)
}
\arguments{
\item{xml_list_obj}{A list object containing XML data.}
}
\value{
A list of attribute values.
}
\description{
This function takes an XML list object and extracts the values of a specific attribute.
}
\keyword{internal}
