% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_search.R
\name{value_grabber}
\alias{value_grabber}
\title{Value Grabber}
\usage{
value_grabber(attribute, xml_list_obj)
}
\arguments{
\item{attribute}{A character string representing the attribute name.}

\item{xml_list_obj}{A list object containing XML data.}
}
\value{
A character string representing the text value of the attribute.
}
\description{
This function takes an attribute and an XML list object, and returns the text value associated with the attribute.
}
\keyword{internal}
